+++ internal/usecase/interfaces.go
@@ -12,6 +12,11 @@ type CommitFetcher interface {
 	FetchCommits(context context.Context, username string, startTime, endTime time.Time) ([]*domain.Commit, error)
 }
 
+// IssueFetcher defines the interface for fetching user issues.
+type IssueFetcher interface {
+	FetchIssueActivities(context context.Context, username string, startTime, endTime time.Time) ([]*domain.Issue, error)
+}
+
 // PullRequestFetcher defines the interface for fetching user pull requests.
 type PullRequestFetcher interface {
 	FetchPullRequests(context context.Context, username string, startTime, endTime time.Time) ([]*domain.PullRequest, error)